+++ PoundPupLegacy.Convert/LocationMigratorPPL.cs
@@ -1384,7 +1384,7 @@ internal sealed class LocationMigratorPPL : MigratorPPL
                 City = GetCity(id, reader.IsDBNull("city") ? null : reader.GetString("city")),
                 PostalCode = GetPostalCode(id, reader.IsDBNull("postal_code") ? null : reader.GetString("postal_code")),
                 SubdivisionId = await GetSubdivisionId(id, subDivisionId, countryId, code, nodeIdReader, subdivisionIdReaderByIso3166Code),
-                CountryId = await GetCountryId(id, countryId, nodeIdReader),
+                CountryId = (int)await GetCountryId(id, countryId, nodeIdReader),
                 Latitude = GetLatitude(id, reader.IsDBNull("latitude") ? null : reader.GetDecimal("latitude")),
                 Longitude = GetLongitude(id, reader.IsDBNull("longitude") ? null : reader.GetDecimal("longitude")),
                 Locatables = new List<LocationLocatable> { new LocationLocatable { LocationId = id, LocatableId = locatableId } }
